feat(): added menu section and separate file for separate canvas experiments
