+++ index.html
@@ -5,11 +5,29 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + TS</title>
+    <link rel="stylesheet" href="./src/style.css" />
+    <link rel="stylesheet" href="./src/root.css" />
+    <link rel="stylesheet" href="./src/main-section.css" />
   </head>
   <body>
-    <!-- <div id="app"></div> -->
-    <canvas id="my-canvas"></canvas>
+    <section>
+      <h1>Main Menu</h1>
+      <p>Welcome to Canvas Playground</p>
 
-    <script type="module" src="/src/main.ts"></script>
+      <ul>
+        <li>
+          <a href="/trailing-effect.html">Trailing Effect</a>
+          <span>Move your mouse around to see the effect</span>
+        </li>
+        <li>
+          <a href="/fireworks.html">Fireworks</a>
+          <span>Watch the awesome HTML5 Canvas fireworks</span>
+        </li>
+        <li>
+          <a href="/physics-math.html"> Physics / Math </a>
+          <span>Math is beautiful</span>
+        </li>
+      </ul>
+    </section>
   </body>
 </html>
