+++ src/canvas-style.css
@@ -1,0 +1,8 @@
+#my-canvas {
+  position: absolute;
+  top: 0;
+  left: 0;
+  background-color: black;
+  height: 100vh;
+  width: 100vw;
+}
