+++ src/main-section.css
@@ -1,0 +1,24 @@
+section {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+section > ul {
+  list-style-type: none;
+}
+section > ul > li {
+  cursor: pointer;
+  margin: 1rem 0;
+  display: flex;
+  flex-direction: column;
+}
+section > ul > li > span {
+  font-size: small;
+}
+section > ul > li > a {
+  transition: 500ms cubic-bezier(0.075, 0.82, 0.165, 1);
+}
+section > ul > li > a:hover {
+  transform: translateX(5px) scale(1.05);
+}
