+++ trailing-effect.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Trailing Effect Canvas</title>
+
+    <link rel="stylesheet" href="./src/style.css" />
+    <link rel="stylesheet" href="./src/root.css" />
+    <link rel="stylesheet" href="./src/canvas-style.css" />
+  </head>
+  <body>
+    <canvas id="my-canvas"></canvas>
+
+    <script type="module" src="/src/trailing.ts"></script>
+  </body>
+</html>
